backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Environment
python-dotenv==1.0.0

# Code Analysis
radon==6.0.1
bandit==1.7.7

# Graph Analysis
networkx==3.2.1

# Git Integration
gitpython==3.1.41

# PDF Generation
reportlab==4.1.0

# LLM Client (Groq)
groq==0.4.2

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
